requirements.txt
boto3
opencv-python
numpy
pandas
torch
torchvision
timm
Pillow
tqdm
pycocotools
scikit-image
matplotlib
ftfy
regex
einops
ensemble-boxes
transformers
streamlit
groundingdino-py

git+https://github.com/IDEA-Research/GroundingDINO.git

git+https://github.com/openai/CLIP.git

